additional functions
