+++ main.py
@@ -2,8 +2,9 @@ import logging
 import os
 
 from aiogram.dispatcher import Dispatcher
+from aiogram.dispatcher.filters import Text
 from aiogram.utils.executor import start_webhook
-from aiogram import Bot
+from aiogram import Bot, types
 
 
 BOT_TOKEN = os.getenv("TG_BOT_TOKEN")
@@ -31,6 +32,30 @@ async def on_startup(dispatcher):
 
 async def on_shutdown(dispatcher):
     await bot.delete_webhook()
+
+
+@dp.message_handler(commands=["start"])
+async def show_list(message: types.Message) -> None:
+    await bot.send_message(message.from_user.id,
+                           f"Привет {message.from_user.first_name}!\n"
+                           "Я бот-агрегатор новостей! 🤖 \n"
+                           "Пожалуйста, выбери функцию из меню")
+
+
+@dp.message_handler()
+async def message_reader(message: types.Message) -> None:
+    await bot.send_message(message.from_user.id, "К сожалению я не знаю данной команды 🙁 \n"
+                                                 "Для справки введите команду: /help")
+
+
+@dp.message_handler(Text(equals="На главную"))
+async def go_back(message: types.Message) -> None:
+    await bot.send_message(message.from_user.id, "Вы перемещены в главное меню.")
+
+
+@dp.message_handler(Text(equals="Цены активов"))
+async def stocks_cos(message: types.Message) -> None:
+    await bot.send_message(message.from_user.id, "Какие активы вас интересуют?")
 
 
 print(__name__)
